# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Teste",
        "description": "Descrição detalhada da API de Teste",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/activity_types": {
            "get": {
                "tags": [
                    "ActivityTypes"
                ],
                "summary": "Lista todos os tipos de atividades",
                "description": "Retorna uma lista de todos os tipos de atividades",
                "operationId": "f777aadf91ec48e476bfe3dcfb256561",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActivityType"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ActivityTypes"
                ],
                "summary": "Cria um novo tipo de atividade",
                "description": "Cria um novo tipo de atividade com os dados fornecidos",
                "operationId": "bffe0d482a9865185d84d63a03d48be9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitação inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/activity_types/{id}": {
            "get": {
                "tags": [
                    "ActivityTypes"
                ],
                "summary": "Obtém um tipo de atividade",
                "description": "Retorna um tipo de atividade pelo ID",
                "operationId": "bf9ebcd77ca13c68f9123c1f130c39d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de atividade a ser visualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de atividade não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ActivityTypes"
                ],
                "summary": "Atualiza um tipo de atividade existente",
                "description": "Atualiza um tipo de atividade com os dados fornecidos",
                "operationId": "d1a05b030321dabfdfb2ebbaf7ae7420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de atividade a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de atividade não encontrado"
                    },
                    "400": {
                        "description": "Solicitação inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ActivityTypes"
                ],
                "summary": "Exclui um tipo de atividade",
                "description": "Exclui um tipo de atividade pelo ID",
                "operationId": "353ae3a65cdee0207515cdf85c259a25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de atividade a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operação bem-sucedida, sem conteúdo a ser retornado"
                    },
                    "404": {
                        "description": "Tipo de atividade não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um novo usuário",
                "description": "Cria um novo usuário com os dados fornecidos",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitação inválida"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Loga o usuário",
                "description": "Loga o usuário usando o e-mail e senha fornecidos",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Senha errada"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Desloga o usuário",
                "description": "Invalida o token atual do usuário",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sessão expirada com sucesso!"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ActivityType": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID do tipo de atividade",
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Nome do tipo de atividade",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Endereço de email do usuário",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            }
        }
    },
...
}